Document that multi-line parameter values aren't supported
We ran into this today and thought we might try to save other users some
debugging time.

I assume it might be possible to improve the plugin to support
multi-line values, but that's beyond my bash abilities to achieve in an
afternoon and this felt like a good first step.